+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - `filter` 1-D digital filter.
 
 - control system module (part 2):
+  - `step` Simulate continuous time model of a state space model or transfer function.
   - `lsim` Plot simulated time response of dynamic system to arbitrary inputs.
   - `dc2` Convert model from discrete to continuous time.
   - `c2d` Convert model from continuous to discrete time.